+++ app/src/main/java/com/ebata_shota/holdemstacktracker/domain/model/GamePlayer.kt
@@ -2,6 +2,6 @@ package com.ebata_shota.holdemstacktracker.domain.model
 
 data class GamePlayer(
     val id: PlayerId,
-    val stack: Double,
+    val stack: Int,
     val isLeaved: Boolean
 )
