+++ app/src/main/java/com/ebata_shota/holdemstacktracker/domain/model/PlayerBaseState.kt
@@ -3,5 +3,5 @@ package com.ebata_shota.holdemstacktracker.domain.model
 data class PlayerBaseState(
     val id: PlayerId,
     val name: String,
-    val stack: Double
+    val stack: Int
 )
