+++ app/src/main/java/com/ebata_shota/holdemstacktracker/domain/model/Pot.kt
@@ -3,7 +3,7 @@ package com.ebata_shota.holdemstacktracker.domain.model
 data class Pot(
     val id: Long,
     val potNumber: Long,
-    val potSize: Double,
+    val potSize: Int,
     val involvedPlayerIds: List<PlayerId>,
     val isClosed: Boolean
 )
