+++ app/src/main/java/com/ebata_shota/holdemstacktracker/domain/usecase/GetMinRaiseSizeUseCase.kt
@@ -5,6 +5,6 @@ import com.ebata_shota.holdemstacktracker.domain.model.Game
 interface GetMinRaiseSizeUseCase {
     suspend fun invoke(
         game: Game,
-        minBetSize: Double
-    ): Double
+        minBetSize: Int
+    ): Int
 }
