+++ app/src/main/java/com/ebata_shota/holdemstacktracker/domain/usecase/GetPendingBetPerPlayerUseCase.kt
@@ -10,5 +10,5 @@ interface GetPendingBetPerPlayerUseCase {
     suspend fun invoke(
         playerOrder: List<PlayerId>,
         actionStateList: List<BetPhaseAction>
-    ): Map<PlayerId, Double>
+    ): Map<PlayerId, Int>
 }
