+++ app/src/main/java/com/ebata_shota/holdemstacktracker/domain/usecase/GetPendingBetSize.kt
@@ -8,5 +8,5 @@ interface GetPendingBetSize {
         actionList: List<BetPhaseAction>,
         playerOrder: List<PlayerId>,
         playerId: PlayerId,
-    ): Double
+    ): Int
 }
