+++ app/src/test/java/com/ebata_shota/holdemstacktracker/domain/usecase/GetNextGamePlayerListUseCaseImplTest.kt
@@ -38,7 +38,7 @@ class GetNextGamePlayerListUseCaseImplTest {
     }
 
     private fun executeAndAssert(
-        pendingBetPerPlayer: Map<PlayerId, Double> = emptyMap(),
+        pendingBetPerPlayer: Map<PlayerId, Int> = emptyMap(),
         players: Set<GamePlayer>,
         action: BetPhaseAction,
         expected: Set<GamePlayer>,
@@ -60,18 +60,18 @@ class GetNextGamePlayerListUseCaseImplTest {
         // prepare
         every { firebaseAuthRepository.myPlayerIdFlow } returns flowOf(PlayerId("0"))
         val players = setOf(
-            GamePlayer(id = PlayerId("0"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+            GamePlayer(id = PlayerId("0"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
         )
-        val action = BetPhaseAction.Blind(playerId = PlayerId("0"), betSize = 100.0)
+        val action = BetPhaseAction.Blind(playerId = PlayerId("0"), betSize = 100)
         executeAndAssert(
             players = players,
             action = action,
             expected = setOf(
-                GamePlayer(id = PlayerId("0"), stack = 900.0, isLeaved = false),
-                GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-                GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+                GamePlayer(id = PlayerId("0"), stack = 900, isLeaved = false),
+                GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+                GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
             )
         )
     }
@@ -81,18 +81,18 @@ class GetNextGamePlayerListUseCaseImplTest {
         // prepare
         every { firebaseAuthRepository.myPlayerIdFlow } returns flowOf(PlayerId("0"))
         val players = setOf(
-            GamePlayer(id = PlayerId("0"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+            GamePlayer(id = PlayerId("0"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
         )
-        val action = BetPhaseAction.Call(playerId = PlayerId("0"), betSize = 100.0)
+        val action = BetPhaseAction.Call(playerId = PlayerId("0"), betSize = 100)
         executeAndAssert(
             players = players,
             action = action,
             expected = setOf(
-                GamePlayer(id = PlayerId("0"), stack = 900.0, isLeaved = false),
-                GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-                GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+                GamePlayer(id = PlayerId("0"), stack = 900, isLeaved = false),
+                GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+                GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
             )
         )
     }
@@ -102,18 +102,18 @@ class GetNextGamePlayerListUseCaseImplTest {
         // prepare
         every { firebaseAuthRepository.myPlayerIdFlow } returns flowOf(PlayerId("0"))
         val players = setOf(
-            GamePlayer(id = PlayerId("0"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+            GamePlayer(id = PlayerId("0"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
         )
-        val action = BetPhaseAction.Raise(playerId = PlayerId("0"), betSize = 400.0)
+        val action = BetPhaseAction.Raise(playerId = PlayerId("0"), betSize = 400)
         executeAndAssert(
             players = players,
             action = action,
             expected = setOf(
-                GamePlayer(id = PlayerId("0"), stack = 600.0, isLeaved = false),
-                GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-                GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+                GamePlayer(id = PlayerId("0"), stack = 600, isLeaved = false),
+                GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+                GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
             )
         )
     }
@@ -123,18 +123,18 @@ class GetNextGamePlayerListUseCaseImplTest {
         // prepare
         every { firebaseAuthRepository.myPlayerIdFlow } returns flowOf(PlayerId("0"))
         val players = setOf(
-            GamePlayer(id = PlayerId("0"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+            GamePlayer(id = PlayerId("0"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
         )
-        val action = BetPhaseAction.Bet(playerId = PlayerId("0"), betSize = 100.0)
+        val action = BetPhaseAction.Bet(playerId = PlayerId("0"), betSize = 100)
         executeAndAssert(
             players = players,
             action = action,
             expected = setOf(
-                GamePlayer(id = PlayerId("0"), stack = 900.0, isLeaved = false),
-                GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-                GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+                GamePlayer(id = PlayerId("0"), stack = 900, isLeaved = false),
+                GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+                GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
             )
         )
     }
@@ -144,18 +144,18 @@ class GetNextGamePlayerListUseCaseImplTest {
         // prepare
         every { firebaseAuthRepository.myPlayerIdFlow } returns flowOf(PlayerId("0"))
         val players = setOf(
-            GamePlayer(id = PlayerId("0"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+            GamePlayer(id = PlayerId("0"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
         )
-        val action = BetPhaseAction.AllIn(playerId = PlayerId("0"), betSize = 1000.0)
+        val action = BetPhaseAction.AllIn(playerId = PlayerId("0"), betSize = 1000)
         executeAndAssert(
             players = players,
             action = action,
             expected = setOf(
-                GamePlayer(id = PlayerId("0"), stack = 0.0, isLeaved = false),
-                GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-                GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+                GamePlayer(id = PlayerId("0"), stack = 0, isLeaved = false),
+                GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+                GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
             )
         )
     }
@@ -165,18 +165,18 @@ class GetNextGamePlayerListUseCaseImplTest {
         // prepare
         every { firebaseAuthRepository.myPlayerIdFlow } returns flowOf(PlayerId("0"))
         val players = setOf(
-            GamePlayer(id = PlayerId("0"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+            GamePlayer(id = PlayerId("0"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
         )
         val action = BetPhaseAction.Check(playerId = PlayerId("0"))
         executeAndAssert(
             players = players,
             action = action,
             expected = setOf(
-                GamePlayer(id = PlayerId("0"), stack = 1000.0, isLeaved = false),
-                GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-                GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+                GamePlayer(id = PlayerId("0"), stack = 1000, isLeaved = false),
+                GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+                GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
             )
         )
     }
@@ -186,18 +186,18 @@ class GetNextGamePlayerListUseCaseImplTest {
         // prepare
         every { firebaseAuthRepository.myPlayerIdFlow } returns flowOf(PlayerId("0"))
         val players = setOf(
-            GamePlayer(id = PlayerId("0"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+            GamePlayer(id = PlayerId("0"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
         )
         val action = BetPhaseAction.Fold(playerId = PlayerId("0"))
         executeAndAssert(
             players = players,
             action = action,
             expected = setOf(
-                GamePlayer(id = PlayerId("0"), stack = 1000.0, isLeaved = false),
-                GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-                GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+                GamePlayer(id = PlayerId("0"), stack = 1000, isLeaved = false),
+                GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+                GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
             )
         )
     }
@@ -207,18 +207,18 @@ class GetNextGamePlayerListUseCaseImplTest {
         // prepare
         every { firebaseAuthRepository.myPlayerIdFlow } returns flowOf(PlayerId("0"))
         val players = setOf(
-            GamePlayer(id = PlayerId("0"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-            GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+            GamePlayer(id = PlayerId("0"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+            GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
         )
         val action = BetPhaseAction.FoldSkip(playerId = PlayerId("0"))
         executeAndAssert(
             players = players,
             action = action,
             expected = setOf(
-                GamePlayer(id = PlayerId("0"), stack = 1000.0, isLeaved = false),
-                GamePlayer(id = PlayerId("1"), stack = 1000.0, isLeaved = false),
-                GamePlayer(id = PlayerId("2"), stack = 1000.0, isLeaved = false),
+                GamePlayer(id = PlayerId("0"), stack = 1000, isLeaved = false),
+                GamePlayer(id = PlayerId("1"), stack = 1000, isLeaved = false),
+                GamePlayer(id = PlayerId("2"), stack = 1000, isLeaved = false),
             )
         )
     }
@@ -228,24 +228,24 @@ class GetNextGamePlayerListUseCaseImplTest {
         // prepare
         every { firebaseAuthRepository.myPlayerIdFlow } returns flowOf(PlayerId("0"))
         val pendingBetPerPlayer = mapOf(
-            PlayerId("0") to 100.0,
-            PlayerId("1") to 200.0,
-            PlayerId("2") to 200.0,
+            PlayerId("0") to 100,
+            PlayerId("1") to 200,
+            PlayerId("2") to 200,
         )
         val players = setOf(
-            GamePlayer(id = PlayerId("0"), stack = 900.0, isLeaved = false),
-            GamePlayer(id = PlayerId("1"), stack = 800.0, isLeaved = false),
-            GamePlayer(id = PlayerId("2"), stack = 800.0, isLeaved = false),
+            GamePlayer(id = PlayerId("0"), stack = 900, isLeaved = false),
+            GamePlayer(id = PlayerId("1"), stack = 800, isLeaved = false),
+            GamePlayer(id = PlayerId("2"), stack = 800, isLeaved = false),
         )
-        val action = BetPhaseAction.Call(playerId = PlayerId("0"), betSize = 200.0)
+        val action = BetPhaseAction.Call(playerId = PlayerId("0"), betSize = 200)
         executeAndAssert(
             pendingBetPerPlayer = pendingBetPerPlayer,
             players = players,
             action = action,
             expected = setOf(
-                GamePlayer(id = PlayerId("0"), stack = 800.0, isLeaved = false),
-                GamePlayer(id = PlayerId("1"), stack = 800.0, isLeaved = false),
-                GamePlayer(id = PlayerId("2"), stack = 800.0, isLeaved = false),
+                GamePlayer(id = PlayerId("0"), stack = 800, isLeaved = false),
+                GamePlayer(id = PlayerId("1"), stack = 800, isLeaved = false),
+                GamePlayer(id = PlayerId("2"), stack = 800, isLeaved = false),
             )
         )
     }
@@ -255,24 +255,24 @@ class GetNextGamePlayerListUseCaseImplTest {
         // prepare
         every { firebaseAuthRepository.myPlayerIdFlow } returns flowOf(PlayerId("0"))
         val pendingBetPerPlayer = mapOf(
-            PlayerId("0") to 100.0,
-            PlayerId("1") to 200.0,
-            PlayerId("2") to 200.0,
+            PlayerId("0") to 100,
+            PlayerId("1") to 200,
+            PlayerId("2") to 200,
         )
         val players = setOf(
-            GamePlayer(id = PlayerId("0"), stack = 900.0, isLeaved = false),
-            GamePlayer(id = PlayerId("1"), stack = 800.0, isLeaved = false),
-            GamePlayer(id = PlayerId("2"), stack = 800.0, isLeaved = false),
+            GamePlayer(id = PlayerId("0"), stack = 900, isLeaved = false),
+            GamePlayer(id = PlayerId("1"), stack = 800, isLeaved = false),
+            GamePlayer(id = PlayerId("2"), stack = 800, isLeaved = false),
         )
-        val action = BetPhaseAction.AllIn(playerId = PlayerId("0"), betSize = 1000.0)
+        val action = BetPhaseAction.AllIn(playerId = PlayerId("0"), betSize = 1000)
         executeAndAssert(
             pendingBetPerPlayer = pendingBetPerPlayer,
             players = players,
             action = action,
             expected = setOf(
-                GamePlayer(id = PlayerId("0"), stack = 0.0, isLeaved = false),
-                GamePlayer(id = PlayerId("1"), stack = 800.0, isLeaved = false),
-                GamePlayer(id = PlayerId("2"), stack = 800.0, isLeaved = false),
+                GamePlayer(id = PlayerId("0"), stack = 0, isLeaved = false),
+                GamePlayer(id = PlayerId("1"), stack = 800, isLeaved = false),
+                GamePlayer(id = PlayerId("2"), stack = 800, isLeaved = false),
             )
         )
     }
@@ -281,21 +281,21 @@ class GetNextGamePlayerListUseCaseImplTest {
     fun action_SB_Check_and_BB_Bet() {
         // prepare
         every { firebaseAuthRepository.myPlayerIdFlow } returns flowOf(PlayerId("1"))
-        val pendingBetPerPlayer = emptyMap<PlayerId, Double>()
+        val pendingBetPerPlayer = emptyMap<PlayerId, Int>()
         val players = setOf(
-            GamePlayer(id = PlayerId("0"), stack = 800.0, isLeaved = false),
-            GamePlayer(id = PlayerId("1"), stack = 800.0, isLeaved = false),
-            GamePlayer(id = PlayerId("2"), stack = 800.0, isLeaved = false),
+            GamePlayer(id = PlayerId("0"), stack = 800, isLeaved = false),
+            GamePlayer(id = PlayerId("1"), stack = 800, isLeaved = false),
+            GamePlayer(id = PlayerId("2"), stack = 800, isLeaved = false),
         )
-        val action = BetPhaseAction.Bet(playerId = PlayerId("1"), betSize = 200.0)
+        val action = BetPhaseAction.Bet(playerId = PlayerId("1"), betSize = 200)
         executeAndAssert(
             pendingBetPerPlayer = pendingBetPerPlayer,
             players = players,
             action = action,
             expected = setOf(
-                GamePlayer(id = PlayerId("0"), stack = 800.0, isLeaved = false),
-                GamePlayer(id = PlayerId("1"), stack = 600.0, isLeaved = false),
-                GamePlayer(id = PlayerId("2"), stack = 800.0, isLeaved = false),
+                GamePlayer(id = PlayerId("0"), stack = 800, isLeaved = false),
+                GamePlayer(id = PlayerId("1"), stack = 600, isLeaved = false),
+                GamePlayer(id = PlayerId("2"), stack = 800, isLeaved = false),
             )
         )
     }
